+++ readMemInfo.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <string_view>
+
+#include <cstdint>
+
+std::uint64_t readMemInfo(const std::string_view &s);
